# Conflicts:
#	pom.xml
#	src/main/java/com/example/usm/controller/ServicesController.java
#	src/main/java/com/example/usm/entity/User.java
#	src/main/java/com/example/usm/repository/ServiceRepository.java
#	src/main/java/com/example/usm/service/services/ServicesService.java
#	src/main/java/com/example/usm/service/user/UserService.java

+++ src/main/java/com/example/usm/service/services/IServicesService.java
@@ -12,7 +12,7 @@ public interface IServicesService {
 
     List<Service> getAll();
 
-    Service findByUID(long uid);
+    Service findByUID(int uid);
 
     List<Service> findByVendor(String vendor);
